chore: docs reorg + Sentry dev-gate + serializable earnings txn + prisma de-bloat (#921)
* docs: relocate scattered root markdown into docs/ + drop scratch files

Move root-level docs to their proper homes (kebab-case, banded docs/ tree):
- TESTING_PLAN.md -> docs/testing/booking-finance-hardening-test-plan.md
- SCHEMA_MAP.md -> docs/prisma/schema-map.md
- CONCURRENCY_AND_LOAD_REPORT.md -> docs/architecture/concurrency-and-load-report.md
- PRODUCT_FLOW_MAP.md -> docs/architecture/product-flow-map.md
- familiarise-demo-tour.md -> docs/guides/app-demo-tour.md
- the three perf investigations -> docs/performance/
Fix the SCHEMA_MAP reference in the enterprise-tests shared setup. Drop scratch
(a.log, deno.lock, stray .txt notes, testsprite tmp log).


* fix(sentry): don't send events from local `next dev` to the shared prod project

All 6 unresolved Sentry issues were development-environment noise from this
session's local E2E testing (environment=development, localhost, streamtest915
on the dev worktrees) — not production errors. Gate `enabled` on
isNotDevelopmentEnvironment() in all three runtimes so a developer's .env DSN
no longer floods the prod project; production/preview (NODE_ENV=production)
capture is unchanged.


* fix(referrals): Serializable isolation + P2034 retry on createEarningsFromPayment (#896 item a)

Deferred from #896's merge (re-indent cascade). The waiver-eligibility count()
ran at default isolation, so two concurrent first-payments for the same new
consultant could both read <3 prior sessions and both waive the platform
commission. Wrap the txn in withSerializableRetry + isolationLevel Serializable
(same pattern as processQualifyingAction/checkout); the P2002 idempotency catch
is preserved outside. Large line count is the prettier re-indent — git diff -w
shows the only logical change is the wrapper. 108 payments/referrals tests green.


* refactor(prisma): extract the BigInt money map to lib/prisma-extensions.ts

lib/prisma.ts went 294 -> 116 lines: the ~190-line #780/#781 result-extension
map (the real bloat) now lives in lib/prisma-extensions.ts, leaving prisma.ts as
just the connection/timeout/client setup. Zero behaviour change — Db/Tx types are
still derived from the composed client (tsc --noEmit clean, no excessive-stack-depth).
Point the #780 drift test at the new file.


* fix(review): skip false-alarm logging for retryable P2034 in the ledger catch + doc nits

#921 review (Gemini HIGH / CodeRabbit MAJOR): the inner postLedgerTxn catch now
sits under withSerializableRetry, so a P2034 serialization conflict was being
recorded as a ledger failure (Sentry + recordSystemError) on every retry. Guard
that logging to non-retryable errors; P2034 just rethrows and retries cleanly.
Also: drop the now-contradictory 'local-only' note in the demo tour (SeedPass123!
is the standard documented seed cred) and add markdownlint H1s to the perf docs.


---------
