+++ .mcp.json.example
@@ -60,6 +60,10 @@
     "betterstack": {
       "type": "http",
       "url": "https://mcp.betterstack.com"
+    },
+    "sentry": {
+      "type": "http",
+      "url": "https://mcp.sentry.dev/mcp"
     }
   }
 }
