+++ instrumentation-client.ts
@@ -3,14 +3,15 @@
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
 import * as Sentry from "@sentry/nextjs";
-import { isProductionEnvironment } from "@/utils/env";
+import { isNotDevelopmentEnvironment, isProductionEnvironment } from "@/utils/env";
 
 const dsn = process.env.NEXT_PUBLIC_SENTRY_DSN;
 
 Sentry.init({
   dsn,
-  // No DSN => Sentry fully disabled (nothing leaves the app); unconfigured envs no-op.
-  enabled: Boolean(dsn),
+  // No DSN => Sentry fully disabled. Also gated off in local `next dev` so a
+  // developer's .env DSN doesn't flood the shared prod project with dev noise.
+  enabled: Boolean(dsn) && isNotDevelopmentEnvironment(),
   environment: process.env.NEXT_PUBLIC_SENTRY_ENVIRONMENT,
 
   // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
